+++ PWGCF/Flow/Tasks/flowJSPCAnalysisTask.cxx
@@ -92,7 +92,7 @@ struct flowJSPCAnalysisTask {
     Configurable<bool> cfgUseCCDB{"cfgUseCCDB", true, "Use CCDB for NUA/NUE corrections."};
     Configurable<std::string> cfgURL{"cfgURL", "http://alice-ccdb.cern.ch",
                                      "Address of the CCDB to get the NUA/NUE."};
-    Configurable<long> cfgTime{"ccdb-no-later-than",
+    Configurable<int> cfgTime{"ccdb-no-later-than",
                                std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::system_clock::now().time_since_epoch()).count(),
                                "Latest acceptable timestamp of creation for the object."};
   } cfgCCDB;
